+++ src/torchCompactRadius/cppSrc/neighborhoodDynamic.cu
@@ -3,10 +3,10 @@
 
 template<std::size_t dim = 2, typename scalar_t = float>
 __global__ void buildNeighborhoodCudaDispatcher(int32_t numParticles,
-                                                cptr_t<int32_t, 1> neighborOffsets, ptr_t<int32_t, 1> neighborList_i, ptr_t<int32_t, 1> neighborList_j,
-                                                cptr_t<scalar_t, 2> queryPositions, cptr_t<scalar_t, 1> querySupport, int searchRange,
+                                                cptr_t<int64_t, 1> neighborOffsets, ptr_t<int64_t, 1> neighborList_i, ptr_t<int64_t, 1> neighborList_j,
+                                                cptr_t<scalar_t, 2> queryPositions, cptr_t<scalar_t, 1> querySupport, int32_t searchRange,
                                                 cptr_t<scalar_t, 2> sortedPositions, cptr_t<scalar_t, 1> sortedSupport,
-                                                cptr_t<int32_t, 2> hashTable, int hashMapLength,
+                                                cptr_t<int64_t, 2> hashTable, int32_t hashMapLength,
                                                 cptr_t<int64_t, 2> cellTable, cptr_t<int32_t, 1> numCells,
                                                 cptr_t<int32_t, 2> offsets, scalar_t hCell, cptr_t<scalar_t, 1> minDomain, cptr_t<scalar_t, 1> maxDomain, cptr_t<int32_t, 1> periodicity,
                                                 supportMode searchMode) {
@@ -17,10 +17,10 @@ __global__ void buildNeighborhoodCudaDispatcher(int32_t numParticles,
 }
 template<std::size_t dim = 2, typename scalar_t = float>
 __global__ void countNeighborsForParticleCudaDispatcher(int32_t numParticles,
-                                                        ptr_t<int32_t, 1> neighborCounters,
-                                                        cptr_t<scalar_t, 2> queryPositions, cptr_t<scalar_t, 1> querySupport, int searchRange,
+                                                        ptr_t<int64_t, 1> neighborCounters,
+                                                        cptr_t<scalar_t, 2> queryPositions, cptr_t<scalar_t, 1> querySupport, int32_t searchRange,
                                                         cptr_t<scalar_t, 2> sortedPositions, cptr_t<scalar_t, 1> sortedSupport,
-                                                        cptr_t<int32_t, 2> hashTable, int hashMapLength,
+                                                        cptr_t<int64_t, 2> hashTable, int32_t hashMapLength,
                                                         cptr_t<int64_t, 2> cellTable, cptr_t<int32_t, 1> numCellsVec,
                                                         cptr_t<int32_t, 2> offsets,
                                                         scalar_t hCell, cptr_t<scalar_t, 1> minDomain, cptr_t<scalar_t, 1> maxDomain, cptr_t<int32_t, 1> periodicity,
@@ -44,10 +44,10 @@ void cuda_error_check() {
 }
 
 template<typename Func, typename... Ts>
-void launchKernel(Func kernel, int numParticles, Ts&&... args) {
-    int blockSize;  // Number of threads per block
-    int minGridSize;  // Minimum number of blocks required for the kernel
-    int gridSize;  // Number of blocks to use
+void launchKernel(Func kernel, int32_t numParticles, Ts&&... args) {
+    int32_t blockSize;  // Number of threads per block
+    int32_t minGridSize;  // Minimum number of blocks required for the kernel
+    int32_t gridSize;  // Number of blocks to use
 
     // Compute the maximum potential block size for the kernel
     cudaOccupancyMaxPotentialBlockSize(&minGridSize, &blockSize, kernel);
@@ -60,9 +60,9 @@ void launchKernel(Func kernel, int numParticles, Ts&&... args) {
 
 
 void buildNeighborhoodCuda(torch::Tensor neighborOffsets, torch::Tensor neighborList_i, torch::Tensor neighborList_j,
-    torch::Tensor queryPositions, torch::Tensor querySupport, int searchRange,
+    torch::Tensor queryPositions, torch::Tensor querySupport, int32_t searchRange,
     torch::Tensor sortedPositions, torch::Tensor sortedSupport,
-    torch::Tensor hashTable, int hashMapLength,
+    torch::Tensor hashTable, int32_t hashMapLength,
     torch::Tensor cellTable, torch::Tensor numCells,
     torch::Tensor offsets, float hCell, torch::Tensor minDomain, torch::Tensor maxDomain, torch::Tensor periodicity,
     supportMode searchMode) {
@@ -72,10 +72,10 @@ void buildNeighborhoodCuda(torch::Tensor neighborOffsets, torch::Tensor neighbor
     int32_t blocks = (int32_t)floor(numParticles / threads) + (numParticles % threads == 0 ? 0 : 1);
 
 #define args numParticles, \
-neighborOffsets.packed_accessor32<int32_t,1, traits>(), neighborList_i.packed_accessor32<int32_t,1, traits>(), neighborList_j.packed_accessor32<int32_t,1, traits>(), \
+neighborOffsets.packed_accessor32<int64_t,1, traits>(), neighborList_i.packed_accessor32<int64_t,1, traits>(), neighborList_j.packed_accessor32<int64_t,1, traits>(), \
 queryPositions.packed_accessor32<scalar_t, 2, traits>(), querySupport.packed_accessor32<scalar_t,1, traits>(), searchRange, \
 sortedPositions.packed_accessor32<scalar_t, 2, traits>(), sortedSupport.packed_accessor32<scalar_t,1, traits>(), \
-hashTable.packed_accessor32<int32_t,2, traits>(), hashMapLength, \
+hashTable.packed_accessor32<int64_t,2, traits>(), hashMapLength, \
 cellTable.packed_accessor32<int64_t,2, traits>(), numCells.packed_accessor32<int32_t,1, traits>(), \
 offsets.packed_accessor32<int32_t,2, traits>(), \
 hCell, minDomain.packed_accessor32<scalar_t,1, traits>(), maxDomain.packed_accessor32<scalar_t,1, traits>(), periodicity.packed_accessor32<int32_t,1, traits>(), searchMode
@@ -105,9 +105,9 @@ hCell, minDomain.packed_accessor32<scalar_t,1, traits>(), maxDomain.packed_acces
 
 void countNeighborsForParticleCuda(
     torch::Tensor neighborCounters, 
-    torch::Tensor queryPositions, torch::Tensor querySupport, int searchRange, 
+    torch::Tensor queryPositions, torch::Tensor querySupport, int32_t searchRange, 
     torch::Tensor sortedPositions, torch::Tensor sortedSupport,
-    torch::Tensor hashTable, int hashMapLength,
+    torch::Tensor hashTable, int32_t hashMapLength,
     torch::Tensor cellTable, torch::Tensor numCellsVec, 
     torch::Tensor offsets,
     float hCell, torch::Tensor minDomain, torch::Tensor maxDomain, torch::Tensor periodicity,
@@ -116,10 +116,10 @@ void countNeighborsForParticleCuda(
 
 #define args \
         numParticles, \
-        neighborCounters.packed_accessor32<int32_t,1, traits>(), \
+        neighborCounters.packed_accessor32<int64_t,1, traits>(), \
         queryPositions.packed_accessor32<scalar_t,2, traits>(), querySupport.packed_accessor32<scalar_t,1, traits>(), searchRange, \
         sortedPositions.packed_accessor32<scalar_t,2, traits>(), sortedSupport.packed_accessor32<scalar_t,1, traits>(), \
-        hashTable.packed_accessor32<int32_t,2, traits>(), hashMapLength, \
+        hashTable.packed_accessor32<int64_t,2, traits>(), hashMapLength, \
         cellTable.packed_accessor32<int64_t,2, traits>(), numCellsVec.packed_accessor32<int32_t,1, traits>(), \
         offsets.packed_accessor32<int32_t,2, traits>(), \
         hCell, minDomain.packed_accessor32<scalar_t, 1, traits>(), maxDomain.packed_accessor32<scalar_t, 1, traits>(), periodicity.packed_accessor32<int32_t, 1, traits>(), searchMode
